LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/browsing_context_name_cross_origin-expected.txt

Blocked access to external URL http://www.localhost:8800/html/browsers/browsing-the-web/history-traversal/browsing_context_name-2.html

page load
Step 1
iframe onload
Step 2
iframe onload
Step 3

FAIL Restoring window.name on cross-origin history traversal assert_equals: expected 2 but got 1
